# Speed up JSON parsing.