By switching to `json.RawMessage` over `map[string]interface{}`, we can
speed up JSON parsing (as we did for nsq_to_redis).

```
 $ benchcmp old.txt new.txt
 benchmark            old ns/op     new ns/op     delta
 BenchmarkServe-8     1601          1485          -7.25%
```